API documented

+++ onyx-backend/budget/src/Budget.Application/Subcategories/RemoveSubcategory/RemoveSubcategoryCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using Abstractions.Messaging;
 using Budget.Domain.Categories;
 using Budget.Domain.Subcategories;
+using Budget.Domain.Transactions;
 using Models.Responses;
 
 namespace Budget.Application.Subcategories.RemoveSubcategory;
@@ -9,11 +10,13 @@ internal sealed class RemoveSubcategoryCommandHandler : ICommandHandler<RemoveSu
 {
     private readonly ISubcategoryRepository _subcategoryRepository;
     private readonly ICategoryRepository _categoryRepository;
+    private readonly ITransactionRepository _transactionRepository;
 
-    public RemoveSubcategoryCommandHandler(ISubcategoryRepository subcategoryRepository, ICategoryRepository categoryRepository)
+    public RemoveSubcategoryCommandHandler(ISubcategoryRepository subcategoryRepository, ICategoryRepository categoryRepository, ITransactionRepository transactionRepository)
     {
         _subcategoryRepository = subcategoryRepository;
         _categoryRepository = categoryRepository;
+        _transactionRepository = transactionRepository;
     }
 
     public async Task<Result> Handle(RemoveSubcategoryCommand request, CancellationToken cancellationToken)
@@ -47,11 +50,32 @@ internal sealed class RemoveSubcategoryCommandHandler : ICommandHandler<RemoveSu
             return Result.Failure(categoryRemoveSubcategoryResult.Error);
         }
 
+        var relatedTransactionsGetResult = await _transactionRepository.GetWhereAsync(
+            t => t.SubcategoryId == subcategoryId,
+            cancellationToken);
+
+        if (relatedTransactionsGetResult.IsFailure)
+        {
+            return Result.Failure(relatedTransactionsGetResult.Error);
+        }
+
+        var relatedTransactions = relatedTransactionsGetResult.Value.ToList();
+
+        relatedTransactions.ForEach(t => t.RemoveSubcategory());
+
         var categoryUpdateResult = await _categoryRepository.UpdateAsync(category, cancellationToken);
 
         if (categoryUpdateResult.IsFailure)
         {
             return Result.Failure(categoryUpdateResult.Error);
+        }
+
+        var transactionsUpdateResult =
+            await _transactionRepository.UpdateRangeAsync(relatedTransactions, cancellationToken);
+
+        if (transactionsUpdateResult.IsFailure)
+        {
+            return Result.Failure(transactionsUpdateResult.Error);
         }
 
         var subcategoryRemoveResult = await _subcategoryRepository.RemoveAsync(subcategory.Id, cancellationToken);
